Добавил статистику

+++ Lab1Hammurabi/Round/RoundSharedInfo.h
@@ -4,7 +4,7 @@ class RoundSharedInfo
 {
 public:
     int StarvedCitizenCount;
-    int StarvedCitizenPercentage;
+    float StarvedCitizenPercentage;
     int AcreWheatIncome;
     int AcrePerCitizenCount;
 };
